+++ adapters/repos/db/vector/hnsw/config_test.go
@@ -95,7 +95,7 @@ func validConfig() Config {
 		RootPath:              "some path",
 		ID:                    "someid",
 		MakeCommitLoggerThunk: func() (CommitLogger, error) { return nil, nil },
-		VectorForIDThunk:      func(context.Context, int32) ([]float32, error) { return nil, nil },
+		VectorForIDThunk:      func(context.Context, uint64) ([]float32, error) { return nil, nil },
 		EFConstruction:        17,
 		MaximumConnections:    50,
 		DistanceProvider:      distancer.NewCosineProvider(),
